Moving the operations routine to a mixin
